+++ src/components/ProgramEditor/WorkoutExerciseSetEditor.tsx
@@ -1,0 +1,10 @@
+export type WorkoutExerciseSetFormData = {
+  weight?: number;
+  reps?: number;
+  time?: number;
+  distance?: number;
+};
+
+export function WorkoutExerciseSetEditor() {
+  return <></>;
+}
